Minor refactoring/renaming
